Move prismic.js into lib folder
