+++ components/PrismicLink.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import {Link as LinkHelper} from 'prismic-reactjs'
-import { linkResolver } from './prismic'
+import { linkResolver } from '../lib/prismic'
 import { Link } from '../routes'
 
 export default (props) => {
