+++ pages/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import NotFound from './notfound'
-import { Client, linkResolver } from '../components/prismic'
+import { Client, linkResolver } from '../lib/prismic'
 import { RichText } from 'prismic-reactjs'
 import { CTABanner, FeaturedItems, NumerotedItems, Separator, TextBlock } from '../components/slices'
 import Layout from './layout'
