+++ pages/products.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from '../routes'
 import NotFound from './notfound'
-import { Client, Prismic, linkResolver } from '../components/prismic'
+import { Client, Prismic, linkResolver } from '../lib/prismic'
 import Layout from './layout'
 import { RichText } from 'prismic-reactjs'
